divide sidebar from layout
