+++ AlexanderMandrov/React/Messenger/src/components/Sidebar/index.js
@@ -1,0 +1,3 @@
+import Sidebar from './Sidebar';
+
+export default Sidebar;
